# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SurveySubmission {
  /** High school GPA */
  gpa: string;
  /** U.S. state the student is from */
  state: string;
  /** Sports played in high school */
  sports: string[];
  /** Favorite Iowa City restaurants */
  restaurants: string[];
  /** Other restaurant if "Other" was selected */
  other_restaurant?: string | null;
}

export interface SurveyResponse {
  id: number;
  created_at: string;
  gpa: string;
  state: string;
  sports: string[];
  restaurants: string[];
  other_restaurant?: string | null;
}

export interface GpaRange {
  range: string;
  count: number;
}

export interface SportCount {
  sport: string;
  count: number;
}

export interface RestaurantCount {
  restaurant: string;
  count: number;
}

export interface StateCount {
  state: string;
  count: number;
  percentage: number;
}

export interface SurveyResults {
  total_responses: number;
  gpa_distribution: GpaRange[];
  sports_counts: SportCount[];
  restaurant_counts: RestaurantCount[];
  top_states: StateCount[];
}

export interface ErrorResponse {
  error: string;
  details?: string | null;
}