# lib/api-zod/src/generated/types/surveyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */

export interface SurveyResponse {
  id: number;
  created_at: Date;
  gpa: string;
  state: string;
  sports: string[];
  restaurants: string[];
  other_restaurant?: string | null;
}