# lib/api-zod/src/generated/types/surveyResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */
import type { GpaRange } from "./gpaRange";
import type { RestaurantCount } from "./restaurantCount";
import type { SportCount } from "./sportCount";
import type { StateCount } from "./stateCount";

export interface SurveyResults {
  total_responses: number;
  gpa_distribution: GpaRange[];
  sports_counts: SportCount[];
  restaurant_counts: RestaurantCount[];
  top_states: StateCount[];
}