# lib/api-zod/src/generated/types/surveySubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */

export interface SurveySubmission {
  /** High school GPA */
  gpa: string;
  /** U.S. state the student is from */
  state: string;
  /** Sports played in high school */
  sports: string[];
  /** Favorite Iowa City restaurants */
  restaurants: string[];
  /** Other restaurant if "Other" was selected */
  other_restaurant?: string | null;
}